+++ lib/attesto_phoenix/controller/backchannel_authentication_controller.ex
@@ -30,7 +30,7 @@ defmodule AttestoPhoenix.Controller.BackchannelAuthenticationController do
 
   @spec create(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def create(conn, params) do
-    config = Config.resolve!()
+    config = Config.resolve!(conn)
     conn = OAuthError.no_store(conn, config)
 
     with :ok <- require_enabled(config),