+++ lib/attesto_phoenix/controller/credential_controller.ex
@@ -40,7 +40,7 @@ defmodule AttestoPhoenix.Controller.CredentialController do
   """
   @spec create(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def create(conn, params) do
-    config = Config.resolve!()
+    config = Config.resolve!(conn)
     resource_metadata = Config.resource_metadata_url(config, conn)
 
     case ProtectedResource.authenticate(conn, config, resource_metadata) do