+++ lib/attesto_phoenix/controller/introspection_controller.ex
@@ -68,7 +68,7 @@ defmodule AttestoPhoenix.Controller.IntrospectionController do
   """
   @spec create(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def create(conn, params) when is_map(params) do
-    config = Config.resolve!()
+    config = Config.resolve!(conn)
     conn = OAuthError.no_store(conn, config)
 
     with :ok <- check_https(conn, config),