+++ lib/attesto_phoenix/controller/token_controller.ex
@@ -88,7 +88,7 @@ defmodule AttestoPhoenix.Controller.TokenController do
   """
   @spec create(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def create(conn, params) do
-    config = Config.resolve!()
+    config = Config.resolve!(conn)
     conn = OAuthError.no_store(conn, config)
 
     with :ok <- require_token_content_type(conn),